@@ -22,7 +22,7 @@ namespace Kudu.Core.Deployment.Oryx
         {
             public static readonly string Node = "8.15";
             public static readonly string Python = "3.6";
-            public static readonly string Dotnet = "2.2";
+            public static readonly string Dotnet = "3.1";
             public static readonly string PHP = "7.3";
         }
 
