@@ -28,9 +28,9 @@ namespace Kudu.Tests.Core.Deployment.Oryx
 
         [Theory]
         [InlineData("NODE", "8.15", true, false, BuildOptimizationsFlags.CompressModules)]
-        [InlineData("PYTHON", "3.6", true, false, BuildOptimizationsFlags.None)]
+        [InlineData("PYTHON", "3.6", true, false, BuildOptimizationsFlags.CompressModules)]
         [InlineData("PHP", "7.3", true, false, BuildOptimizationsFlags.None)]
-        [InlineData("DOTNETCORE", "2.2", true, true, BuildOptimizationsFlags.UseTempDirectory)]
+        [InlineData("DOTNETCORE", "3.1", true, true, BuildOptimizationsFlags.UseTempDirectory)]
         public void ArgumentPropertyTest(string language, string version,
             bool expectedRunOryxBuild, bool expectedSkipKuduSync, BuildOptimizationsFlags expectedFlags)
         {
@@ -51,13 +51,13 @@ namespace Kudu.Tests.Core.Deployment.Oryx
 
         [Theory]
         [InlineData("NODE", "8.12",
-            "oryx build OutputPath -o OutputPath --platform nodejs --platform-version 8.12 -i BuildTempPath -p compress_node_modules=tar-gz")]
+            "oryx build RepositoryPath -o BuildTempPath --platform nodejs -i BuildTempPath -p compress_node_modules=tar-gz")]
         [InlineData("PYTHON", "3.6",
-            "oryx build OutputPath -o OutputPath --platform python --platform-version 3.6 -p virtualenv_name=antenv3.6")]
+            "oryx build RepositoryPath -o BuildTempPath --platform python --platform-version 3.6 -i BuildTempPath -p compress_virtualenv=tar-gz -p virtualenv_name=antenv3.6")]
         [InlineData("PHP", "7.3",
-            "oryx build OutputPath -o OutputPath --platform php --platform-version 7.3")]
-        [InlineData("DOTNETCORE", "2.2",
-            "oryx build RepositoryPath -o OutputPath --platform dotnet --platform-version 2.2 -i BuildTempPath")]
+            "oryx build OutputPath -o OutputPath --platform php")]
+        [InlineData("DOTNETCORE", "3.1",
+            "oryx build RepositoryPath -o OutputPath --platform dotnet --platform-version 3.1 -i BuildTempPath")]
         public void CommandGenerationTest(string language, string version, string expectedCommand)
         {
             var mockedEnvironment = new Dictionary<string, string>()
@@ -128,15 +128,15 @@ namespace Kudu.Tests.Core.Deployment.Oryx
             {
                 IOryxArguments args = new AppServiceOryxArguments(TestMockedEnvironment.GetMockedEnvironment());
                 string command = args.GenerateOryxBuildCommand(mockedContext, TestMockedEnvironment.GetMockedEnvironment());
-                Assert.Equal("oryx build OutputPath -o OutputPath --platform nodejs --platform-version 8.12 -i BuildTempPath -p compress_node_modules=zip",
+                Assert.Equal("oryx build RepositoryPath -o BuildTempPath --platform nodejs -i BuildTempPath",
                     args.GenerateOryxBuildCommand(mockedContext, TestMockedEnvironment.GetMockedEnvironment()));
             }
         }
 
         [Theory]
-        [InlineData("2.7", "oryx build OutputPath -o OutputPath --platform python --platform-version 2.7 -p virtualenv_name=antenv2.7")]
-        [InlineData("3.6", "oryx build OutputPath -o OutputPath --platform python --platform-version 3.6 -p virtualenv_name=antenv3.6")]
-        [InlineData("3.7", "oryx build OutputPath -o OutputPath --platform python --platform-version 3.7 -p virtualenv_name=antenv")]
+        [InlineData("2.7", "oryx build RepositoryPath -o BuildTempPath --platform python --platform-version 2.7 -i BuildTempPath -p compress_virtualenv=tar-gz -p virtualenv_name=antenv2.7")]
+        [InlineData("3.6", "oryx build RepositoryPath -o BuildTempPath --platform python --platform-version 3.6 -i BuildTempPath -p compress_virtualenv=tar-gz -p virtualenv_name=antenv3.6")]
+        [InlineData("3.7", "oryx build RepositoryPath -o BuildTempPath --platform python --platform-version 3.7 -i BuildTempPath -p compress_virtualenv=tar-gz -p virtualenv_name=antenv")]
         public void PythonVirtualEnvironmentNamingTest(string version, string expectedCommand)
         {
             var mockedEnvironment = new Dictionary<string, string>()
@@ -231,7 +231,7 @@ namespace Kudu.Tests.Core.Deployment.Oryx
             {
                 IOryxArguments args = new AppServiceOryxArguments(TestMockedEnvironment.GetMockedEnvironment());
                 string command = args.GenerateOryxBuildCommand(mockedContext, TestMockedEnvironment.GetMockedEnvironment());
-                Assert.Equal("oryx build  -o  --platform python --platform-version 3.6 -p virtualenv_name=antenv3.6",
+                Assert.Equal("oryx build RepositoryPath -o BuildTempPath --platform python --platform-version 3.6 -i BuildTempPath -p compress_virtualenv=tar-gz -p virtualenv_name=antenv3.6",
                     args.GenerateOryxBuildCommand(mockedContext, TestMockedEnvironment.GetMockedEnvironment()));
             }
         }
