@@ -8,6 +8,7 @@ namespace Kudu.Tests.Core.Deployment.Oryx
     [Collection("MockedEnvironmentVariablesCollection")]
     public class OryxArgumentsLinuxConsumptionFunctionAppTests
     {
+        // TODO check if it is proper change
         [Fact]
         public void DefaultTest()
         {
@@ -23,7 +24,7 @@ namespace Kudu.Tests.Core.Deployment.Oryx
                 OutputPath = "OutputPath"
             };
             string command = args.GenerateOryxBuildCommand(mockedContext, TestMockedEnvironment.GetMockedEnvironment());
-            Assert.Equal("oryx build RepositoryPath -o RepositoryPath", command);
+            Assert.Equal("oryx build RepositoryPath -o OutputPath", command);
         }
 
         [Theory]
@@ -48,16 +49,16 @@ namespace Kudu.Tests.Core.Deployment.Oryx
             }
         }
 
+        // TODO Check if this is the proper change.
         [Theory]
         [InlineData("NODE",
-            "oryx build RepositoryPath -o RepositoryPath --platform nodejs --platform-version 8.15")]
+            "oryx build RepositoryPath -o OutputPath --platform nodejs --platform-version 8.15")]
         [InlineData("PYTHON",
-            "oryx build RepositoryPath -o RepositoryPath --platform python --platform-version 3.6 " +
-            "-p packagedir=.python_packages\\lib\\python3.6\\site-packages")]
+            "oryx build RepositoryPath -o OutputPath --platform python --platform-version 3.6 -p packagedir=.python_packages\\lib\\python3.6\\site-packages")]
         [InlineData("PHP",
-            "oryx build RepositoryPath -o RepositoryPath --platform php --platform-version 7.3")]
+            "oryx build RepositoryPath -o OutputPath --platform php --platform-version 7.3")]
         [InlineData("DOTNET",
-            "oryx build RepositoryPath -o RepositoryPath --platform dotnet --platform-version 2.2")]
+            "oryx build RepositoryPath -o OutputPath --platform dotnet --platform-version 3.1")]
         public void CommandGenerationTest(string functions_worker_runtime, string expected_command)
         {
             var mockedEnvironment = new Dictionary<string, string>()
